refactoring command
