+++ backend/command/src/main/kotlin/CommandConfiguration.kt
@@ -1,0 +1,57 @@
+package de.holisticon.ranked.command
+
+import de.holisticon.ranked.extension.DefaultSmartLifecycle
+import mu.KLogging
+import org.axonframework.commandhandling.SimpleCommandBus
+import org.axonframework.config.EventHandlingConfiguration
+import org.axonframework.messaging.interceptors.BeanValidationInterceptor
+import org.springframework.beans.factory.annotation.Autowired
+import org.springframework.context.annotation.Bean
+import org.springframework.context.annotation.Configuration
+import org.springframework.validation.beanvalidation.LocalValidatorFactoryBean
+import javax.annotation.PostConstruct
+import javax.validation.ValidatorFactory
+
+/**
+ * This is the main spring configuration class for everything axon/command related.
+ */
+@Configuration
+class CommandConfiguration(val eventHandlingConfiguration: EventHandlingConfiguration) {
+
+  companion object : KLogging() {
+    const val PHASE_REPLAY = Int.MAX_VALUE - 10
+  }
+
+  @PostConstruct
+  fun init() {
+    eventHandlingConfiguration.processors.map {
+      logger.info { "Processor: ${it.name} " }
+    }
+
+    logger.info { "processors: ${eventHandlingConfiguration.processors} " }
+  }
+
+  /**
+   * Enable bean validation for incoming commands.
+   */
+  @Autowired
+  fun configure(bus: SimpleCommandBus, validationFactory: ValidatorFactory) {
+    bus.registerDispatchInterceptor(BeanValidationInterceptor(validationFactory))
+  }
+
+  /**
+   * Provide bean validation validator.
+   */
+  @Bean
+  fun validatorFactoryBean(): ValidatorFactory = LocalValidatorFactoryBean()
+
+  /**
+   * Register Lifecycle handler for event replay.
+   */
+  @Bean
+  fun replayTrackingProcessors() = object : DefaultSmartLifecycle(PHASE_REPLAY) {
+    override fun onStart() {
+      logger.info { "requesting replay" }
+    }
+  }
+}
