+++ backend/command/src/main/kotlin/handler/LoggingEventHandler.kt
@@ -1,13 +1,14 @@
-package de.holisticon.ranked.command.logging
+package de.holisticon.ranked.command.handler
 
-import de.holisticon.ranked.axon.TrackingProcessor
 import de.holisticon.ranked.model.event.MatchCreated
 import mu.KLogging
 import org.axonframework.config.ProcessingGroup
 import org.axonframework.eventhandling.EventHandler
 import org.springframework.stereotype.Component
 
-@TrackingProcessor
+/**
+ * Tracking [EventHandler] just to keep the log.
+ */
 @Component
 @ProcessingGroup("Logging")
 class LoggingEventHandler {
@@ -16,7 +17,7 @@ class LoggingEventHandler {
 
   @EventHandler
   fun on(event: MatchCreated) {
-    // logger.trace { "Match created ${event.matchId}" }
+    logger.trace { "Match created ${event.matchId}" }
   }
 
 }
