+++ backend/command/src/test/kotlin/saga/EloSagaSpec.kt
@@ -2,7 +2,7 @@ package de.holisticon.ranked.command.saga
 
 import de.holisticon.ranked.command.api.ParticipateInMatch
 import de.holisticon.ranked.command.api.UpdatePlayerRanking
-import de.holisticon.ranked.command.service.EloCalculationService
+import de.holisticon.ranked.command.elo.EloCalculationService
 import de.holisticon.ranked.model.MatchSet
 import de.holisticon.ranked.model.Team
 import de.holisticon.ranked.model.UserName
