+++ backend/view/player/src/main/kotlin/PlayerView.kt
@@ -1,6 +1,5 @@
 package de.holisticon.ranked.view.player
 
-import de.holisticon.ranked.axon.TrackingProcessor
 import de.holisticon.ranked.model.UserName
 import de.holisticon.ranked.model.event.PlayerCreated
 import de.holisticon.ranked.model.event.PlayerRankingChanged
@@ -9,8 +8,6 @@ import org.axonframework.config.ProcessingGroup
 import org.axonframework.eventhandling.EventHandler
 import org.springframework.stereotype.Service
 
-
-@TrackingProcessor
 @ProcessingGroup(PlayerViewService.NAME)
 @Service
 class PlayerViewService {
