+++ backend/view/pom.xml
@@ -23,11 +23,12 @@
       <groupId>${project.groupId}</groupId>
       <artifactId>ranked-model</artifactId>
     </dependency>
-    <dependency>
-      <groupId>${project.groupId}</groupId>
-      <artifactId>ranked-axon</artifactId>
-    </dependency>
 
+    <dependency>
+      <groupId>org.axonframework</groupId>
+      <artifactId>axon-core</artifactId>
+      <version>${axon.version}</version>
+    </dependency>
     <dependency>
       <groupId>org.springframework</groupId>
       <artifactId>spring-web</artifactId>
