+++ backend/view/wall/src/main/kotlin/WallView.kt
@@ -1,6 +1,5 @@
 package de.holisticon.ranked.view.wall
 
-import de.holisticon.ranked.axon.TrackingProcessor
 import de.holisticon.ranked.model.MatchSet
 import de.holisticon.ranked.model.Team
 import de.holisticon.ranked.model.UserName
@@ -10,19 +9,17 @@ import io.swagger.annotations.ApiOperation
 import mu.KLogging
 import org.axonframework.config.ProcessingGroup
 import org.axonframework.eventhandling.EventHandler
-import org.springframework.context.ApplicationEventPublisher
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RestController
 import java.time.LocalDateTime
 
 
-@TrackingProcessor
 @ProcessingGroup(WallView.NAME)
-@Api(tags = arrayOf("News wall"))
+@Api(tags = ["News wall"])
 @RestController
 @RequestMapping(value = "/view")
-class WallView(val eventPublisher: ApplicationEventPublisher) {
+class WallView() {
 
   companion object : KLogging() {
     const val NAME = "Wall"
@@ -88,9 +85,6 @@ class WallView(val eventPublisher: ApplicationEventPublisher) {
   fun on(e: PlayerParticipatedInMatch) {
     logger.info { "Player ${e.player} with ranking ${e.eloRanking} played in match ${e.matchId}" }
   }
-
-//  @GetMapping("/management/replay")
-//  fun replay() = eventPublisher.publishEvent(ReplayTrackingProcessor(NAME))
 }
 
 data class Match(
