+++ extensions/README.md
@@ -4,3 +4,5 @@ Collection of (not ranked related) extensions to standard spring+java.
 
 Separate from the main project because these could be useful in other places, might be a good
 idea to move them to a toolisticon module.
+
+TODO: make toolisticon-extensions
